requirements.txt
# SPDX-FileCopyrightText: 2021 ladyada for Adafruit Industries
#
# SPDX-License-Identifier: Unlicense

Adafruit-Blinka
adafruit-circuitpython-busdevice
adafruit-circuitpython-register
adafruit-circuitpython-pca9685
adafruit-circuitpython-motor
python-socketio
aiohttp
Adafruit_PCA9685